Enhance formula generation logic

+++ Benchmark/scriptsS5/main.py
@@ -1,8 +1,9 @@
 import sys
 import random
+import os
 import s5
 
-benchmark_size = 36
+benchmark_size = 4
 benchmark_instance = 0
 benchmark_path = "../formulasS5/"
 
@@ -29,24 +30,27 @@ if __name__ == '__main__':
 
     elif len(argv) == 1 and argv[0] == "-benchmark" :
         l = 3
-        p = 0.5
         # in this case, we generate the benchmark based on the relationship between clauses and the number of variables
         
         n_values = [60, 80, 100, 120]
         
-        for i in range(2,benchmark_size) :
-            
-            for n in n_values:
-                # j is the ratio of clauses to variables
-                j = 1
-                while j <= 15.0 :
-                    m = int(n*j) # 4,3 is the threshold for 3-SAT problems, we can use it as a reference for S5 problems
-                    fname = benchmark_path+f"formula{i}-{n}-{m}-{l}-{p}.s5"
-                    print("Generating formula: "+fname)
-                    ffile = open(fname,'w')
-                    ffile.write(f"{s5.phi(n, m, l, p)}")
-                    ffile.close()
-                    j+=(0.5)
+        for p in [0.25, 0.5, 0.75, 1.0] :
+            for i in range(2,benchmark_size) :
+                
+                for n in n_values:
+                    # j is the ratio of clauses to variables
+                    j = 1
+                    while j <= 25.0 :
+                        m = int(n*j) # 4,3 is the threshold for 3-SAT problems, we can use it as a reference for S5 problems
+                        fname = benchmark_path+f"formula{i}-{n}-{m}-{l}-{p}.s5"
+                        if os.path.exists(fname) :
+                            print("Skipping (ya existe): "+fname)
+                        else :
+                            print("Generating formula: "+fname)
+                            ffile = open(fname,'w')
+                            ffile.write(f"{s5.phi(n, m, l, p)}")
+                            ffile.close()
+                        j+=(0.5)
     else:
         print(f'error: incorrect number of arguments.')
         print(f'  use: {argv[0]} | -benchmark')
